+++ app/src/organisms/RunDetails/index.tsx
@@ -1,6 +1,18 @@
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
-import { Flex, Text, DIRECTION_COLUMN } from '@opentrons/components'
+import {
+  Flex,
+  PrimaryBtn,
+  Text,
+  BORDER_WIDTH_DEFAULT,
+  C_BLUE,
+  C_WHITE,
+  DIRECTION_COLUMN,
+  FONT_WEIGHT_SEMIBOLD,
+  LINE_HEIGHT_SOLID,
+  SPACING_2,
+  SPACING_3,
+} from '@opentrons/components'
 import { Page } from '../../atoms/Page'
 import { useProtocolDetails } from './hooks'
 
@@ -8,9 +20,25 @@ export function RunDetails(): JSX.Element | null {
   const { t } = useTranslation('run_details')
   const { displayName, protocolData } = useProtocolDetails()
   if (protocolData == null) return null
+  const cancelRunButton = (
+    <PrimaryBtn
+      onClick={() => console.log('Cancel Run')}
+      backgroundColor={C_WHITE}
+      color={C_BLUE}
+      borderWidth={BORDER_WIDTH_DEFAULT}
+      lineHeight={LINE_HEIGHT_SOLID}
+      fontWeight={FONT_WEIGHT_SEMIBOLD}
+      marginX={SPACING_3}
+      paddingRight={SPACING_2}
+      paddingLeft={SPACING_2}
+    >
+      Cancel Run
+    </PrimaryBtn>
+  )
 
   const titleBarProps = {
     title: t('protocol_title', { protocol_name: displayName }),
+    rightNode: cancelRunButton,
   }
 
   return (
